arreglado int a float64

# Conflicts:
#	proyectoNovedades.sh

+++ servicios/novedades/novedades.go
@@ -92,7 +92,7 @@ type Distribuciones struct {
 }
 
 type RecursosNovedades struct {
-	Importe     int           `bson:"importe"`
+	Importe     float64       `bson:"importe"`
 	Comentarios string        `bson:"comentarios"`
 	Recurso     string        `bson:"recurso"`
 	Periodo     string        `bson:"periodo"`
